Maps/requirements.txt
Django==1.8.5
django-extensions==1.7.4
django-geojson==2.8.1
django-leaflet==0.16.0
django-wms==0.1.12
djangorestframework==3.4.7
psycopg2==2.6.1
six==1.10.0
